bug fix in filter (single result in nan)

+++ spikeSort/highPassFilter.m
@@ -12,7 +12,7 @@ function [xf_detect, xf, noise_std_detect, noise_std_sorted, thr, thrmax] = high
     %HIGH-PASS FILTER OF THE DATA
     if exist('ellip', 'file')                               % Checks for the signal processing toolbox
         [b_detect,a_detect] = ellip(param.detect_order, 0.1, 40, [fmin_detect fmax_detect]*2/sr);
-        xf_detect = filtfilt(b_detect, a_detect, x);
+        xf_detect = filtfilt(b_detect, a_detect, double(x));
         if fmin_sort ~= fmin_detect || fmax_sort ~= fmax_detect || param.sort_order ~= param.detect_order
             [b,a] = ellip(param.sort_order, 0.1, 40, [fmin_sort fmax_sort]*2/sr);
             xf = filtfilt(b, a, x);
